+++ CHANGES.rst
@@ -4,6 +4,7 @@
 
 0.6.4-dev 
 =========
+  * ``taxit --ranked`` now takes one of two arguments ``{rows, columns}``
   * Records containing the word "group" in taxonomic name are no longer flagged as unclassified/invalid [GH-93]
   
 0.6.4